requirements.txt
google-generativeai>=0.8.0
chromadb>=0.5.0
langgraph>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
tenacity>=8.3.0
pytest>=8.0.0
